requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.0
pymongo>=4.8.0
httpx>=0.27.0
cryptography>=43.0.0
python-multipart>=0.0.9
passlib>=1.7.4

# LangChain / LangGraph
langchain>=0.2.14
langgraph>=0.2.31
langchain-openai>=0.1.19

# Shopify Dev MCP client (stdio → npx @shopify/dev-mcp); requires Python >= 3.10
mcp>=1.2.0

# Tests
pytest>=8.0.0
mongomock>=4.1.0
